third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gatherND.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] gatherND float32 3D input and 2D indices
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherND float32 actual -66.05901336669922 should be close enough to expected 90.2870101928711 by ULP distance: expected a number less than or equal to 0n but got 2235085098n
[FAIL] [required] gatherND float16 3D input and 2D indices
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherND float16 actual -66.0625 should be close enough to expected 90.3125 by ULP distance: expected a number less than or equal to 0 but got 32380
Harness: the test ran to completion.
